+++ setup.cfg
@@ -1,2 +1,3 @@
 [flake8]
-exclude = .eggs,*.egg,build
+exclude = .eggs,*.egg,build,.anaconda*
+ignore = H304
